+++ packages/adminui-vue/components/UTableEntity/filter-templates/string/notIsNull.vue
@@ -4,19 +4,21 @@
 
 <script>
 export default {
-  name: 'FilterEntityIsNotNull',
-
+  name: 'FilterStringNotIsNull',
   components: {
     FilterTemplate: require('../../components/FilterTemplate.vue').default
   },
 
   mixins: [require('../mixinForFilter.js')],
 
+  created () {
+    this.$emit('search-disabled', false)
+  },
   methods: {
     getCondition () {
       return {
-        whereList: [{ condition: 'isNotNull' }],
-        description: 'isNotNull'
+        whereList: [{ condition: 'notIsNull' }],
+        description: 'notIsNull'
       }
     }
   }